Fall back to GetSystemTimeAsFileTime() if precise is unavailable.
GetSystemTimePreciseAsFileTime() is only supported from Windows 8
onwards. On Windows 7, only GetSystemTimeAsFileTime() is available.
Sadly, this is not caught by Windows SDK header guards, because of
https://wpdev.uservoice.com/forums/110705/suggestions/35836531. As
a result, FlexASIO currently crashes on startup on Windows 7 due to
the missing DLL import.

This commit fixes the issue by dynamically importing
GetSystemTimePreciseAsFileTime(), falling back to
GetSystemTimeAsFileTime() if that fails.

Fixes #15.