Introduce a TOML configuration file, and add backend PA API option.
With this commit, FlexASIO will attempt to load a configuration file
named "FlexASIO.toml" in the user's profile directory. If found,
it is parsed as TOML. Currently, only one option is supported:
"backend", which can be set to one of the PortAudio host API names
(e.g. "MME") to change the PortAudio host API that FlexASIO uses.

See #6.