By default, use the sample rate from a previous instance, if any.
REW has a bug where it resets the ASIO driver but then doesn't tell
the new instance of the driver which sample rate it's using. To
work around that bug, we use an ugly hack where we retain the sample
rate from a previous instance of the driver using a global variable.
We then use that as the default sample rate.

Fixes #31.