Adjust buffer size and suggested latency defaults, and update docs.
Following experiments with full-duplex operation in REW, it has been
observed that the best way to get glitch-free operation is to use an
ASIO buffer size of at least 20 ms and a suggested latency of at
least 3 times the buffer size. This commit makes these the new
defaults, and updates the documentation with the findings.