Integrate the Inno Setup installer generation with CMake.
This commit completely revamps the way Inno Setup interacts with the
build system. The packaging procedure is now driven by a separate
CMake script (instead of Inno Setup), the build now starts from a
clean slate, and the Inno Setup script is greatly simplified as lots
of logic is moved to CMake.

See #36.