Vendor the PortAudio dependency as a submodule with ExternalProject.
This commits adds a submodule pointing to the portaudio
pa_stable_v190600_20161030 (same as latest FlexASIO release bundle)
and extends the FlexASIO CMake superbuild to build the PortAudio
dependency. This ensures that FlexASIO is built against a
well-defined PortAudio library, and will make it easy to migrate
to a different PortAudio revision in the future. It also gets rid of
the dependency on vcpkg.

Closes #37.