Document the Windows audio buffer size as 10 ms
After some double checking it turns out 20 ms is wrong - the buffer size
that the Windows Audio Engine uses internally is 10 ms (this is the size
of the buffer that APOs are called with, and it's also the "packet" size
it uses with WaveRT hardware).

20 ms may have been correct at some point in the distant past (Windows 7
maybe?), but it's clearly not anymore.